requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
plaid-python==16.0.0
pandas==1.3.3
openpyxl==3.0.9
python-dotenv==0.19.0
psycopg2-binary==2.9.9
pandas==2.1.0
python-dotenv==1.0.0